@@ -1,17 +1,19 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
     <ItemGroup>
       <PackageReference Include="FluentAssertions" Version="5.10.3" />
-      <PackageReference Include="Microsoft.AspNetCore.App" />
-      <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="2.2.0" />
-      <PackageReference Include="Nunit" Version="3.11.0" />
-      <PackageReference Include="NUnit3TestAdapter" Version="3.11.0" />
-      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
+      <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.5" />
+      <PackageReference Include="Nunit" Version="3.12.0" />
+      <PackageReference Include="NUnit3TestAdapter" Version="3.16.1">
+        <PrivateAssets>all</PrivateAssets>
+        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      </PackageReference>
+      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
       <PackageReference Include="SpecFlow" Version="3.3.30" />
       <PackageReference Include="SpecFlow.NUnit" Version="3.3.30" />
       <PackageReference Include="SpecFlow.Tools.MsBuild.Generation" Version="3.3.30" />
