@@ -1,25 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="AspNetCore.HealthChecks.SqlServer" Version="2.2.0" />
-        <PackageReference Include="MediatR" Version="6.0.0" />
-        <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="6.0.1" />
+        <PackageReference Include="AspNetCore.HealthChecks.SqlServer" Version="3.1.1" />
+        <PackageReference Include="MediatR" Version="8.0.2" />
+        <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="8.0.1" />
+        <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.5" />
         <PackageReference Include="Microsoft.AspNetCore.Diagnostics.HealthChecks" Version="2.2.0" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="2.2.3" />
-        <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="2.2.0" />
-        <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.Abstractions" Version="2.2.0" />
-        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.7.1" />
-        <PackageReference Include="Microsoft.AspNetCore.App" />
-        <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-        <PackageReference Include="Microsoft.Extensions.Options" Version="2.2.0" />
-        <PackageReference Include="NLog.Web.AspNetCore" Version="4.8.0" />
-        <PackageReference Include="SFA.DAS.NLog.Targets.Redis" Version="1.1.5" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="3.1.5" />
+        <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="3.1.5" />
+        <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.Abstractions" Version="3.1.5" />
+        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.14.0" />
+        <PackageReference Include="Microsoft.Extensions.Options" Version="3.1.5" />
+        <PackageReference Include="NLog.Web.AspNetCore" Version="4.9.2" />
+        <PackageReference Include="SFA.DAS.NLog.Targets.Redis" Version="1.2.1" />
         <PackageReference Include="SFA.DAS.Configuration.AzureTableStorage" Version="3.0.84" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.2.1" />
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.5.1" />
     </ItemGroup>
 
     <ItemGroup>
@@ -36,6 +35,9 @@
             <CopyToPublishDirectory>Never</CopyToPublishDirectory>
         </Content>
         <Content Update="nlog.config">
+            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </Content>
+        <Content Update="DataFiles\frameworks_1.0.json">
             <CopyToOutputDirectory>Always</CopyToOutputDirectory>
         </Content>
     </ItemGroup>
