@@ -7,12 +7,13 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using Microsoft.OpenApi.Models;
 using SFA.DAS.Configuration.AzureTableStorage;
 using SFA.DAS.Courses.Api.AppStart;
 using SFA.DAS.Courses.Api.Infrastructure;
-using SFA.DAS.Courses.Application.StandardsImport.Handlers.ImportStandards;
+using SFA.DAS.Courses.Application.CoursesImport.Handlers.ImportStandards;
 using SFA.DAS.Courses.Domain.Configuration;
 using SFA.DAS.Courses.Domain.Interfaces;
 using SFA.DAS.Courses.Infrastructure.HealthCheck;
@@ -28,14 +29,18 @@ namespace SFA.DAS.Courses.Api
             var config = new ConfigurationBuilder()
                 .AddConfiguration(configuration)
                 .SetBasePath(Directory.GetCurrentDirectory())
-#if DEBUG
-                .AddJsonFile("appsettings.json", true)
-                .AddJsonFile("appsettings.Development.json", true)
-#endif
                 .AddEnvironmentVariables();
-
+            
+            
             if (!configuration["Environment"].Equals("DEV", StringComparison.CurrentCultureIgnoreCase))
             {
+                
+#if DEBUG
+                config
+                    .AddJsonFile("appsettings.json", true)
+                    .AddJsonFile("appsettings.Development.json", true);
+#endif
+                
                 config.AddAzureTableStorage(options =>
                     {
                         options.ConfigurationKeys = configuration["ConfigNames"].Split(",");
@@ -58,26 +63,32 @@ namespace SFA.DAS.Courses.Api
             services.Configure<AzureActiveDirectoryConfiguration>(_configuration.GetSection("AzureAd"));
             services.AddSingleton(cfg => cfg.GetService<IOptions<AzureActiveDirectoryConfiguration>>().Value);
             
-            var serviceProvider = services.BuildServiceProvider();
-
+            var coursesConfiguration = _configuration
+                .GetSection("Courses")
+                .Get<CoursesConfiguration>();
+            
             if (!ConfigurationIsLocalOrDev())
             {
-                services.AddAuthentication(serviceProvider.GetService<IOptions<AzureActiveDirectoryConfiguration>>().Value);
+                var azureAdConfiguration = _configuration
+                    .GetSection("AzureAd")
+                    .Get<AzureActiveDirectoryConfiguration>();
 
+                services.AddAuthentication(azureAdConfiguration);
             }
 
-            var coursesConfiguration = serviceProvider.GetService<IOptions<CoursesConfiguration>>().Value;
-            services.AddHealthChecks()
-                .AddSqlServer(coursesConfiguration.ConnectionString)
-                .AddCheck<LarsHealthCheck>("Lars Data Health Check",
-                    failureStatus: HealthStatus.Unhealthy,
-                    tags: new[] {"ready"})
-                .AddCheck<InstituteOfApprenticeshipServiceHealthCheck>("IFATE Health Check",
-                    failureStatus: HealthStatus.Unhealthy,
-                    tags: new[] {"ready"});
+            if (_configuration["Environment"] != "DEV")
+            {
+                services.AddHealthChecks()
+                    .AddSqlServer(coursesConfiguration.ConnectionString)
+                    .AddCheck<LarsHealthCheck>("Lars Data Health Check",
+                        failureStatus: HealthStatus.Unhealthy,
+                        tags: new[] {"ready"})
+                    .AddCheck<InstituteOfApprenticeshipServiceHealthCheck>("IFATE Health Check",
+                        failureStatus: HealthStatus.Unhealthy,
+                        tags: new[] {"ready"});
+            }
             
-
-            services.AddMediatR(typeof(ImportStandardsCommand).Assembly);
+            services.AddMediatR(typeof(ImportDataCommand).Assembly);
 
             services.AddServiceRegistration();
 
@@ -90,7 +101,7 @@ namespace SFA.DAS.Courses.Api
                     {
                         o.Conventions.Add(new AuthorizeControllerModelConvention());
                     }
-                }).SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                }).SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
             
             services.AddApplicationInsightsTelemetry(_configuration["APPINSIGHTS_INSTRUMENTATIONKEY"]);
 
@@ -99,36 +110,38 @@ namespace SFA.DAS.Courses.Api
                 c.SwaggerDoc("v1", new OpenApiInfo { Title = "CoursesAPI", Version = "v1" });
             });
             
-            serviceProvider = services.BuildServiceProvider();
-            var indexBuilder = serviceProvider.GetService<IIndexBuilder>();
-            indexBuilder.Build();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IIndexBuilder indexBuilder)
         {
+            indexBuilder.Build();
+            
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
             
-            app.UseAuthentication();    
-        
-            app.UseHealthChecks();
-            
-            app.UseMvc(routes =>
+            app.UseAuthentication();
+
+            if (!_configuration["Environment"].Equals("DEV", StringComparison.CurrentCultureIgnoreCase))
             {
-                routes.MapRoute(
-                    name: "default",
-                    template: "api/{controller=Standards}/{action=Index}/{id?}");
-            });
+                app.UseHealthChecks();    
+            }
             
+            app.UseRouting();
+            app.UseEndpoints(builder =>
+            {
+                builder.MapControllerRoute(
+                    name: "default",
+                    pattern: "api/{controller=Standards}/{action=Index}/{id?}");
+            });
+
             app.UseSwagger();
             app.UseSwaggerUI(c =>
             {
                 c.SwaggerEndpoint("/swagger/v1/swagger.json", "CoursesAPI");
                 c.RoutePrefix = string.Empty;
             });
-            
         }
         
         private bool ConfigurationIsLocalOrDev()
