@@ -1,5 +1,4 @@
 using System.Linq;
-using Microsoft.EntityFrameworkCore.Internal;
 
 namespace SFA.DAS.Courses.Domain.Entities
 {
@@ -10,16 +9,16 @@ namespace SFA.DAS.Courses.Domain.Entities
             return new StandardImport
             {
                 Id = standard.LarsCode,
-                CoreSkillsCount = standard.Duties.Any() ? standard.Duties.Where(c=>c.IsThisACoreDuty.Equals(1)).Select(c=>c.DutyDetail).Join("|") : null,
+                CoreSkillsCount = standard.Duties.Any() ? string.Join("|", standard.Duties.Where(c=>c.IsThisACoreDuty.Equals(1)).Select(c=>c.DutyDetail)) : null,
                 IntegratedDegree = standard.IntegratedDegree,
                 Level = standard.Level,
                 OverviewOfRole = standard.OverviewOfRole,
                 StandardPageUrl = standard.StandardPageUrl.AbsoluteUri,
                 Title = standard.Title,
                 TypicalDuration = standard.TypicalDuration,
-                TypicalJobTitles = standard.TypicalJobTitles.Join("|"),
+                TypicalJobTitles = string.Join("|", standard.TypicalJobTitles),
                 Version = standard.Version,
-                Keywords = standard.Keywords.Any() ? standard.Keywords.Join("|") : null,
+                Keywords = standard.Keywords.Any() ? string.Join("|",standard.Keywords) : null,
                 RouteId = standard.RouteId
             };
         }
